Rename mapping.py to msid_mbid_mapping.py
