+++ listenbrainz/db/model/pinned_recording.py
@@ -4,7 +4,7 @@ from typing import Optional, List
 from pydantic import validator, constr, NonNegativeInt
 
 from data.model.validators import check_datetime_has_tzinfo
-from listenbrainz.db.mapping import MsidMbidModel, load_recordings_from_mapping
+from listenbrainz.db.msid_mbid_mapping import MsidMbidModel, load_recordings_from_mapping
 from listenbrainz.messybrainz import load_recordings_from_msids
 
 DAYS_UNTIL_UNPIN = 7  # default = unpin after one week
